requirements.txt
# Vercel Python runtime + local uvicorn
requests>=2.31.0
Pillow>=10.0.0
fastapi>=0.110.0
uvicorn>=0.30.0
